+++ andropilot/src/main/scala/com/geeksville/gmaps/Scene.scala
@@ -76,7 +76,7 @@ class Scene(val map: GoogleMap) extends AndroidLogger {
   }
 
   def render() {
-    map.clear()
+    //map.clear() // FIXME - don't add this back until we are sure it won't blow away the plane icon
     renderMarkers()
     renderSegments()
   }